@@ -4,44 +4,49 @@ import com.ZharikovaES.PersonalListsApp.services.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityCustomizer;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.web.SecurityFilterChain;
+
 
 @Configuration
 @EnableWebSecurity
-public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
+public class WebSecurityConfig {
     @Autowired
     private UserService userService;
-
-    @Autowired
-    private PasswordEncoder passwordEncoder;
 
     @Bean
     public PasswordEncoder getPasswordEncoder() {
         return new BCryptPasswordEncoder(8);
     }
 
+    @Bean
+    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
 
-    @Override
-    protected void configure(HttpSecurity http) throws Exception {
-        http.csrf(csrf -> csrf.disable());
-        http.authorizeRequests(requests -> requests
-                .antMatchers( "/", "/css/**", "/js/**", "/img/**", "/activate/*").permitAll()
-                .anyRequest().authenticated())
-                .formLogin(login -> login
-                        .loginPage("/login").loginProcessingUrl("/login").defaultSuccessUrl("/home").failureUrl("/login?error")
+        http.csrf((csrf) -> csrf
+                    .ignoringRequestMatchers("/**"))
+                .authorizeHttpRequests((authorize) -> authorize
+                    .requestMatchers("/", "/main", "/registration")
+                    .permitAll()
+                    .anyRequest().authenticated())
+                .formLogin(form -> form
+                        .loginPage("/login")
+                        .loginProcessingUrl("/login")
+                        .defaultSuccessUrl("/home")
+                        .failureUrl("/login?error")
                         .permitAll())
                 .logout(logout -> logout.deleteCookies("JSESSIONID"))
-                .rememberMe(me -> me.key("uniqueAndSecret"));
+                .userDetailsService(userService)
+                .rememberMe(rm -> rm.key("uniqueAndSecret").userDetailsService(userService));
+
+        return http.build();
     }
 
-    @Override
-    protected void configure(AuthenticationManagerBuilder auth) throws Exception {
-        auth.userDetailsService(userService)
-                .passwordEncoder(passwordEncoder);
+    @Bean
+    public WebSecurityCustomizer webSecurityCustomizer() {
+        return web -> web.debug(true).ignoring().requestMatchers("/css/**", "/js/**", "/img/**", "/activate/*", "/favicon.ico");
     }
 }
