@@ -2,7 +2,9 @@ package com.ZharikovaES.PersonalListsApp.models;
 
 import com.fasterxml.jackson.annotation.*;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
+import org.hibernate.annotations.ColumnDefault;
+
 import java.io.Serializable;
 import java.util.*;
 
@@ -14,7 +16,7 @@ public class List extends UnitData implements Serializable {
     @OneToMany(fetch = FetchType.EAGER, mappedBy = "list", cascade = CascadeType.ALL , orphanRemoval = true)
     private Set<Item> items = new HashSet<>();
 
-    @ManyToMany(cascade = { CascadeType.ALL })
+    @ManyToMany(cascade = { CascadeType.PERSIST, CascadeType.MERGE })
     @JoinTable(
             name = "list_tags",
             joinColumns = { @JoinColumn(name = "list_id") },
@@ -23,13 +25,18 @@ public class List extends UnitData implements Serializable {
     @JsonIgnoreProperties("lists")
     private Set<Tag> tags = new HashSet<>();
 
+    @Column(name = "view_order", nullable = false)
+    @ColumnDefault("0")
+    private int order;
+
     public List() {
     }
 
-    public List(String title, Set<Item> items, Set<Tag> tags, User user) {
+    public List(String title, Set<Item> items, Set<Tag> tags, User user, int order) {
         super(user, title);
         this.items = items;
         this.tags = tags;
+        this.order = order;
     }
 
     public Set<Item> getItems() {
@@ -48,4 +55,25 @@ public class List extends UnitData implements Serializable {
         this.tags = tags;
     }
 
+    public int getOrder() {
+        return order;
+    }
+
+    public void setOrder(int order) {
+        this.order = order;
+    }
+
+    @JsonGetter("items")
+    public java.util.List<Item> getItemsList(){
+        return items.stream().sorted(Comparator.comparingInt(Item::getOrder)).toList();
+    }
+
+    @JsonSetter("items")
+    public void setItemsList(java.util.List<Item> items){
+        this.items.clear();
+        for(int i = 0; i < items.size(); i++){
+            items.get(i).setOrder(i);
+        }
+        this.items.addAll(items);
+    }
 }
