@@ -1,6 +1,7 @@
 package com.ZharikovaES.PersonalListsApp.repos;
 
 import com.ZharikovaES.PersonalListsApp.models.List;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
@@ -13,7 +14,7 @@ public interface ListRepo extends CrudRepository<List, Long> {
 
     @Modifying
     @Query("delete from List l where l.id = :id")
-    void deleteById(Long id);
+    void deleteById(@Param("id") Long id);
 
     @Query("select l from List l join l.tags t where t.id = :tId and l.userId = :uId order by l.dateUpdate asc")
     java.util.List<List> findAllByUserIdByTagIdByDateAsc(@Param("uId") Long uId, @Param("tId") Long tId);
@@ -38,6 +39,13 @@ public interface ListRepo extends CrudRepository<List, Long> {
 
     @Query("select l from List l where l.userId = :uId and l.title LIKE CONCAT('%', :title, '%') order by l.dateUpdate asc")
     java.util.List<List> findAllByUserIdByTitleByDateAsc(@Param("uId") Long uId, @Param("title") String title);
+
+    java.util.List<List> findByUserIdAndTitleIsLike(@Param("uId") Long uId, @Param("title") String title, Pageable pageable);
+    @Query("select l from List l join l.tags t where t.id = :tId and l.userId = :uId and l.title LIKE CONCAT('%', :title, '%')")
+    java.util.List<List> findByUserIdAndTagIdAndTitleIsLike(@Param("uId") Long uId, @Param("tId") Long tId, @Param("title") String title, Pageable pageable);
+    java.util.List<List> findByUserId(@Param("uId") Long uId, Pageable pageable);
+    @Query("select l from List l join l.tags t where t.id = :tId and l.userId = :uId")
+    java.util.List<List> findByUserIdAndTagId(@Param("uId") Long uId, @Param("tId") Long tId, Pageable pageable);
 
     java.util.List<List> findAll();
 }
